% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_static.R
\docType{data}
\name{bs_static}
\alias{bs_static}
\title{Static methods for the \code{bspline} class}
\format{
An object of class \code{bspline_static_methods} of length 4.
}
\usage{
bs_static
}
\value{
A list that contains all the static methods for the \link{bspline} class.
}
\description{
Static methods for the \link{bspline} class can be use with the
syntax : \code{bs_static$static_method_name()}.
}
\section{Fields}{

\describe{
\item{\code{point_of_support()}}{This static method assumes that the splines are
terminated at the boundary with repeat knot points.

Arguments:
\itemize{
\item \code{t_knot} a Matrix. Spline knot points.
\item \code{K} an integer. Order of polynomial.
\item \code{D} an array.
}}

\item{\code{pp_coeff_from_spline_coeff()}}{Returns the piecewise polynomial coefficients in matrix C
from spline coefficients in vector m.

Arguments:
\itemize{
\item \code{m} Spline coefficients (\eqn{m = M \cdot D}).
\item \code{t_knot} a two-column matrix. Spline knot points.
\item \code{K} an integer. Order of polynomial.
\item \code{B} An array. Optional
}}

\item{\code{evaluate_from_pp_coeff()}}{Returns the value of the function with derivative \code{D}
represented by \code{PP} coefficients \code{C} at locations \code{t}.
\code{t_pp} containes the intervals.

Arguments:
\itemize{
\item \code{t} a matrix. Points locations.
\item \code{C} a vector. Piecewise polynomial coefficients.
\item \code{t_pp} a matrix. pp break points. (\code{size(t_pp) = nrow(t_knot) - 2*K + 1}).
\item \code{D}
}}

\item{\code{spline()}}{Returns the basis splines of order K evaluated at point t,
given knot points t_knot. If you optionally provide D, then D derivatives
will be returned.
\code{dim(B) = c(nrow(t), M, D)} where \code{M = nrow(t_knot) - K} is the
number of splines.

Arguments:
\itemize{
\item \code{t} a matrix. Points locations.
\item \code{t_knot} a Matrix. Spline knot points.
\item \code{K} an integer. Order of polynomial.
\item \code{D}
}}
}}

\keyword{datasets}
