% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_static.R
\docType{data}
\name{point_of_support}
\alias{point_of_support}
\alias{pp_coeff_from_spline_coeff}
\alias{evaluate_from_pp_coeff}
\alias{spline}
\alias{bs_utils}
\title{Utils for the \code{bspline} class}
\format{
An object of class \code{bspline_static_methods} of length 4.
}
\usage{
point_of_support(t_knot, K, D)

pp_coeff_from_spline_coeff(m, t_knot, K, B)

evaluate_from_pp_coeff(t, C, t_pp, D)

spline(t, t_knot, K, D)

bs_utils
}
\arguments{
\item{t_knot}{a matrix. Spline knot points.}

\item{K}{an integer. Order of polynomial.}

\item{D}{an array.}

\item{m}{a numeric. Spline coefficients (\eqn{m = M \cdot D}).}

\item{B}{an array.}

\item{t}{a matrix. Point locations.}

\item{C}{a vector. Piecewise polynomial coefficients (\code{size(C) = [nrow(t_pp)-1, K]})}

\item{t_pp}{a matrix. pp break points (\code{size(t_pp) = nrow(t_knot) - 2*K + 1}).}
}
\value{
A list that contains functions for the \link{bspline} class.
}
\description{
A list of utils for the \link{bspline} class. They can be
used with the synthax \code{bs_utils$function_name()}. These functions are
equivalent to static methods for the \code{bspline} class.
}
\section{Functions}{
\itemize{
\item \code{point_of_support()}: This static method assumes that the splines are terminated at the boundary with repeat knot points.

\item \code{pp_coeff_from_spline_coeff()}: Returns the piecewise polynomial coefficients in matrix C from spline coefficients in vector m.

\item \code{evaluate_from_pp_coeff()}: Returns the value of the function with derivative \code{D} represented by \code{PP} coefficients \code{C} at locations \code{t}. \code{t_pp} containes the intervals.

\item \code{spline()}: Returns the basis splines of order K evaluated at point t, given knot points t_knot. If you optionally provide D, then D derivatives will be returned. \code{dim(B) = c(nrow(t), M, D)} where \code{M = nrow(t_knot) - K} is the number of splines.

}}
\keyword{datasets}
