% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline.R
\name{bspline}
\alias{bspline}
\title{Create a B-Spline.}
\description{
Ici la description
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-bspline-clone}{\code{bspline$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bspline-clone"></a>}}
\if{latex}{\out{\hypertarget{method-bspline-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bspline$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
