% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_smoothing_spline.R
\name{gps_smoothing_spline}
\alias{gps_smoothing_spline}
\title{gps_smoothing_spline: a class useful for smoothing noisy gps data and removing
outliers.}
\description{
gps_smoothing_spline: a class useful for smoothing noisy gps data and removing
outliers.

gps_smoothing_spline: a class useful for smoothing noisy gps data and removing
outliers.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-gps_smoothing_spline-clone}{\code{gps_smoothing_spline$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gps_smoothing_spline-clone"></a>}}
\if{latex}{\out{\hypertarget{method-gps_smoothing_spline-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gps_smoothing_spline$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
