% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolating_spline.R
\name{smoothing_spline}
\alias{smoothing_spline}
\title{R6 class that interpolates splines.}
\value{
Fitted data
}
\description{
This R6 class inherits from the \code{bspline} class.
}
\section{Super class}{
\code{\link[smouseGPS:bspline]{smouseGPS::bspline}} -> \code{smoothing_spline}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{t}}{array of values for the independent axis.}

\item{\code{x}}{array of values for the dependent axis.}

\item{\code{K}}{Order of the spline.}

\item{\code{fun}}{Cubic spline interpolant.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{K}}{Order of the spline.}

\item{\code{fun}}{Cubic spline interpolant.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-smoothing_spline-new}{\code{smoothing_spline$new()}}
\item \href{#method-smoothing_spline-clone}{\code{smoothing_spline$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="smouseGPS" data-topic="bspline" data-id="["><a href='../../smouseGPS/html/bspline.html#method-bspline-['><code>smouseGPS::bspline$[()</code></a></span></li>
<li><span class="pkg-link" data-pkg="smouseGPS" data-topic="bspline" data-id="[<-"><a href='../../smouseGPS/html/bspline.html#method-bspline-[<-'><code>smouseGPS::bspline$[<-()</code></a></span></li>
<li><span class="pkg-link" data-pkg="smouseGPS" data-topic="bspline" data-id="print"><a href='../../smouseGPS/html/bspline.html#method-bspline-print'><code>smouseGPS::bspline$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="smouseGPS" data-topic="bspline" data-id="set_m"><a href='../../smouseGPS/html/bspline.html#method-bspline-set_m'><code>smouseGPS::bspline$set_m()</code></a></span></li>
<li><span class="pkg-link" data-pkg="smouseGPS" data-topic="bspline" data-id="value_at_points"><a href='../../smouseGPS/html/bspline.html#method-bspline-value_at_points'><code>smouseGPS::bspline$value_at_points()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-smoothing_spline-new"></a>}}
\if{latex}{\out{\hypertarget{method-smoothing_spline-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{smoothing_spline$new(t, x, distribution, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-smoothing_spline-clone"></a>}}
\if{latex}{\out{\hypertarget{method-smoothing_spline-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{smoothing_spline$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
