% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\docType{data}
\name{lambda}
\alias{lambda}
\title{Get lambda methods.}
\format{
An object of class \code{lambda_methods} of length 8.
}
\usage{
lambda
}
\value{
A character which represents the lamba methods.
}
\description{
Get a list of available lambda methods.
}
\keyword{datasets}
