% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smouseGPS-package.R
\docType{package}
\name{smouseGPS-package}
\alias{smouseGPS}
\alias{smouseGPS-package}
\title{smouseGPS: Interpolating splines, smoothing splines and b-splines}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A series of classes for interpolating and smoothing data using b-splines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://alexmerot.github.io/smouseGPS/}
}

}
\author{
\strong{Maintainer}: Alexis Merot \email{alexis.merot@ornitho79.org}

}
\keyword{internal}
