% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bspline.R
\name{bspline}
\alias{bspline}
\title{R6 class that creates b-splines.}
\value{
b-splines
}
\description{
Creates b-splines given some set of knot points, and then
evaluates the splines given some set of coefficients for the splines. It has
3 arguments initialisation : \code{f = BSpline(K, t_knot, m)}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{t_knot}}{a Matrix. Spline knot points.}

\item{\code{K}}{an integer. Order of polynomial.}

\item{\code{m}}{a numeric. Spline coefficients (\eqn{m = M \cdot D}).}

\item{\code{B}}{a numeric array with 3 dimension.}

\item{\code{x_mean}}{a numeric. If set, these will be used to scale the output
(default to 0).}

\item{\code{x_std}}{a numeric. \eqn{x_{out} = x_{std} \cdot (X \cdot m) + x_{mean}}}

\item{\code{t_pp}}{a numeric columnwise matrix. pp break points (\code{size(t_pp) = nrow(t_knot) - 2*K + 1}).}

\item{\code{C}}{a vector. Piecewise polynomial coefficients
(\code{size(C) = [nrow(t_pp)-1, K]}).}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{get_S}}{Get \eqn{S = K - 1}}

\item{\code{get_domain}}{Get the domain of \code{t_knot}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-bspline-new}{\code{bspline$new()}}
\item \href{#method-bspline-value_at_points}{\code{bspline$value_at_points()}}
\item \href{#method-bspline-clone}{\code{bspline$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bspline-new"></a>}}
\if{latex}{\out{\hypertarget{method-bspline-new}{}}}
\subsection{Method \code{new()}}{
Instantiate an object of the class \code{bspline}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bspline$new(t_knot, K, m)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t_knot}}{Spline knot points.}

\item{\code{K}}{Order of polynomial.}

\item{\code{m}}{Spline coefficients (\eqn{m = M \cdot D}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bspline-value_at_points"></a>}}
\if{latex}{\out{\hypertarget{method-bspline-value_at_points}{}}}
\subsection{Method \code{value_at_points()}}{
Get the value at points.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bspline$value_at_points(t, num_derivatives)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{a matrix. Points locations.}

\item{\code{num_derivatives}}{an integer. Numero of the derivatives.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bspline-clone"></a>}}
\if{latex}{\out{\hypertarget{method-bspline-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{bspline$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
