% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_spline.R
\name{smoothing_spline}
\alias{smoothing_spline}
\title{smoothing_spline: Fit noisy data with a tensioned interpolating spline.}
\description{
smoothing_spline: Fit noisy data with a tensioned interpolating spline.

smoothing_spline: Fit noisy data with a tensioned interpolating spline.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-smoothing_spline-clone}{\code{smoothing_spline$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-smoothing_spline-clone"></a>}}
\if{latex}{\out{\hypertarget{method-smoothing_spline-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{smoothing_spline$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
